requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
python-multipart

crewai
crewai[tools]
langchain-openai
openai

twilio

deepface
tensorflow-cpu
opencv-python-headless

python-jose[cryptography]
python-dotenv
requests
Pillow
numpy
sqlalchemy
alembic
structlog
passlib[bcrypt]
aiofiles
httpx